refactor: extract device utilities into a separate header file
